chore: 폴더 아키텍쳐 정리
