+++ packages/ui/src/components/button/style.css.ts
@@ -1,6 +1,6 @@
+import { spacingSprinkles } from "@styles/sprinkles.css";
+import { Color } from "@styles/theme.css";
 import { style, styleVariants } from "@vanilla-extract/css";
-import { spacingSprinkles } from "../globalStyle/sprinkles.css";
-import { Color } from "../globalStyle/theme.css";
 
 const base = style([
   spacingSprinkles({
