+++ packages/ui/src/components/imageUploader/ImageUploader.stories.ts
@@ -1,4 +1,4 @@
-import { Meta, StoryObj } from "@storybook/react";
+import type { Meta, StoryObj } from "@storybook/react";
 import ImageUploader from "./ImageUploader";
 
 const meta: Meta<typeof ImageUploader> = {
