+++ packages/ui/src/styles/ColorPallet/ColorPallet.mdx
