+++ packages/ui/tsconfig.json
@@ -2,7 +2,14 @@
   "extends": "@repo/typescript-config/react-library.json",
   "compilerOptions": {
     "outDir": "dist",
-    "typeRoots": ["./node_modules/@types", "src/ImageUploader/types"]
+    "typeRoots": ["./node_modules/@types", "src/ImageUploader/types"],
+    "paths": {
+      "@/*": ["./src/*"],
+      "@components/*": ["./src/components/*"],
+      "@styles/*": ["./src/styles/*"],
+      "@store/*": ["./src/store/*"],
+      "@assets/*": ["./assets/*"]
+    }
   },
   "include": ["src", "store"],
   "exclude": ["node_modules", "dist"]
